+++ src/test/kotlin/com/ryzingtitan/datalogparser/cucumber/common/DatalogRecordRepositoryStepDefs.kt
@@ -79,8 +79,8 @@ class DatalogRecordRepositoryStepDefs(private val datalogRecordRepository: Datal
         return DatalogRecord(
             sessionId = UUID.fromString(tableRow["sessionId"]),
             timestamp = Instant.parse(tableRow["timestamp"]),
-            longitude = tableRow["longitude"].toString().toFloat(),
-            latitude = tableRow["latitude"].toString().toFloat(),
+            longitude = tableRow["longitude"].toString().toDouble(),
+            latitude = tableRow["latitude"].toString().toDouble(),
             altitude = tableRow["altitude"].toString().toFloat(),
             intakeAirTemperature = tableRow["intakeAirTemperature"].toString().toIntOrNull(),
             boostPressure = tableRow["boostPressure"].toString().toFloatOrNull(),
